refactor(cate-agent): extract host-agnostic CateAgentThread body
